requirements.txt
django
django-environ
django-cors-headers
whitenoise
gunicorn
razorpay
pillow
psycopg2-binary
